refactor: 달력 버튼 아이콘 변경 및 타입, 유틸 함수 분리
