+++ components/atoms/icons/index.ts
@@ -1,8 +1,7 @@
 export * from './camera-icon';
 export * from './check-icon';
-export { CheckInCircleIcon } from './check-in-circle-icon';
-export type { Directions } from './chevron-icon';
-export { ChevronIcon } from './chevron-icon';
+export * from './check-in-circle-icon';
+export * from './chevron-icon';
 export * from './date-arrow-icon';
 export { DownArrowIcon } from './down-arrow-icon';
 export { LeftArrowIcon } from './left-arrow-icon';
@@ -10,3 +9,4 @@ export * from './search-icon';
 export * from './star-icon';
 export * from './star-icon-fill';
 export * from './swim-icon';
+export * from './types';
