+++ components/atoms/icons/types.ts
@@ -1,0 +1,1 @@
+export type Directions = 'left' | 'up' | 'right' | 'down';
